// src/lib/image/types.ts
export interface ImageFilters {
    brightness: number;
    contrast: number;
    saturation: number;
    temperature: number;
    sharpness: number;
    exposure: number;
    highlights: number;
    shadows: number;
    vibrance: number;
    hue: number;
    blur: number;
    sepia: number;
    grayscale: number;
    invert: number;
    noise: number;
    pixelate: number;
    glitchIntensity: number;
    glitchOffset: number;
    dreamEffect: number;
    mosaicSize: number;
    pencilEffect: number;
    watercolorEffect: number;
    textureOpacity: number;
    colorPalette: number;
    vignette: number;
    kaleidoscope: number;
    fractalize: number;
    doubleExposure: number;
    neonGlow: number;
    neonColor: number;
    oilPaint: number;
    popArt: number;
    retroFilm: number;
    retroTint: number;
    anaglyph3d: number;
    anaglyphOffset: number;
    lightLeak: number;
    lightLeakColor: number;
}

export const defaultFilters: ImageFilters = {
    brightness: 0,
    contrast: 0,
    saturation: 0,
    temperature: 0,
    sharpness: 0,
    exposure: 0,
    highlights: 0,
    shadows: 0,
    vibrance: 0,
    hue: 0,
    blur: 0,
    sepia: 0,
    grayscale: 0,
    invert: 0,
    noise: 0,
    pixelate: 0,
    glitchIntensity: 0,
    glitchOffset: 0,
    dreamEffect: 0,
    mosaicSize: 0,
    pencilEffect: 0,
    watercolorEffect: 0,
    textureOpacity: 0,
    colorPalette: 0,
    vignette: 0,
    kaleidoscope: 0,
    fractalize: 0,
    doubleExposure: 0,
    neonGlow: 0,
    neonColor: 0,
    oilPaint: 0,
    popArt: 0,
    retroFilm: 0,
    retroTint: 0,
    anaglyph3d: 0,
    anaglyphOffset: 0,
    lightLeak: 0,
    lightLeakColor: 0
}; 